+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Added methods `EntityPlayerClientComponent.viewPlayerListChange()` and `EntityPlayerClientComponent.viewPlayerPermission()`.
 - (API) Added method `DyeColor.getColor()` to get the corresponded `Color` object of the dye color.
 - (API) Implemented firework and firework star. Multiple classes including `FireworkType`, `FireworkExplosion` were added.
+- Player's height will be changed when the player is sneaking, swimming, gliding, or crawling now.
 
 ### Changed